Move functions to plot.py
